requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.0
scikit-learn==1.4.0
google-generativeai==0.3.2
gunicorn==21.2.0
